tests/e2e/requirements.txt
# Python dependencies for kubectl-mtv e2e tests
pytest>=7.0.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0
pytest-xdist>=2.5.0
pyyaml>=6.0
python-dotenv>=0.19.0
